+++ packages/common-ethereum/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Bump `@subql/common` dependency
+
 ## [4.5.4] - 2024-10-23
 ### Changed
 - Bump `@subql/common` dependency